Add TracerBase new abstraction. Sync to commit 92edaa9.
